// frontend/src/app/api/model/learningCard.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LearningStatus } from './learningStatus';
import { Milestone } from './milestone';
import { Note } from './note';


export interface LearningCard { 
    id: number;
    title: string;
    description?: string | null;
    status: LearningStatus;
    createdAt: string;
    milestones: Array<Milestone>;
    notes: Array<Note>;
    milestonesTotal: number;
    milestonesDone: number;
    progressPercent: number;
}
export namespace LearningCard {
}